requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.16
dj-database-url>=2.1,<3
whitenoise>=6.6,<7
markdown>=3.5,<4
bleach>=6.1,<7
gunicorn>=21.2,<23
psycopg2-binary>=2.9,<3
argon2-cffi>=23.1,<24
